fix bug that xinglong commited

+++ consensus-module/poc/consensus-poc-base/src/main/java/io/nuls/consensus/poc/tx/processor/YellowPunishTxProcessor.java
@@ -36,7 +36,7 @@ public class YellowPunishTxProcessor implements TransactionProcessor<YellowPunis
         YellowPunishData punishData = tx.getTxData();
         List<byte[]> deletedList = new ArrayList<>();
         for (byte[] address : punishData.getAddressList()) {
-            boolean result = storageService.delete(this.getPoKey(address, (byte) PunishType.YELLOW.getCode(), tx.getBlockHeight()));
+            boolean result = storageService.delete(this.getPoKey(address, PunishType.YELLOW.getCode(), tx.getBlockHeight()));
             if (!result) {
                 BlockHeader header = (BlockHeader) secondaryData;
                 BlockRoundData roundData = new BlockRoundData(header.getExtend());
@@ -73,7 +73,7 @@ public class YellowPunishTxProcessor implements TransactionProcessor<YellowPunis
             boolean result = storageService.save(po);
             if (!result) {
                 for (byte[] bytes : savedList) {
-                    this.storageService.delete(getPoKey(bytes, (byte) PunishType.YELLOW.getCode(), header.getHeight()));
+                    this.storageService.delete(getPoKey(bytes, PunishType.YELLOW.getCode(), header.getHeight()));
                 }
                 throw new NulsRuntimeException(KernelErrorCode.FAILED, "rollback tx failed!");
             } else {
